refactor: WaitingController 분리
