+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -13,7 +13,6 @@ import org.springframework.test.context.jdbc.Sql;
 import roomescape.controller.request.AdminReservationRequest;
 import roomescape.controller.request.MemberLoginRequest;
 import roomescape.controller.request.ReservationRequest;
-import roomescape.controller.request.WaitingRequest;
 
 import java.lang.reflect.Field;
 import java.time.LocalDate;
@@ -176,62 +175,5 @@ class ReservationControllerTest {
                 .when().get("/reservations-mine")
                 .then().log().all()
                 .statusCode(200);
-    }
-
-    @DisplayName("예약 대기를 추가할 수 있다.")
-    @Test
-    void should_add_waiting_when_request() {
-        MemberLoginRequest loginRequest = new MemberLoginRequest("1234", "sun@email.com");
-
-        String cookie = RestAssured
-                .given().log().all()
-                .contentType(ContentType.JSON)
-                .body(loginRequest)
-                .when().post("/login")
-                .then().statusCode(200)
-                .extract().header("Set-Cookie");
-
-        WaitingRequest request = new WaitingRequest(
-                LocalDate.of(2030, 8, 5), 6L, 10L);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(request)
-                .cookie(cookie)
-                .when().post("/waiting")
-                .then().log().all()
-                .statusCode(201)
-                .header("Location", "/waiting/1");
-    }
-
-    @DisplayName("존재하는 예약 대기라면 예약 대기를 삭제할 수 있다.")
-    @Test
-    void should_delete_waiting_when_waiting_exist() {
-        MemberLoginRequest loginRequest = new MemberLoginRequest("1234", "sun@email.com");
-
-        String cookie = RestAssured
-                .given().log().all()
-                .contentType(ContentType.JSON)
-                .body(loginRequest)
-                .when().post("/login")
-                .then().statusCode(200)
-                .extract().header("Set-Cookie");
-
-        WaitingRequest request = new WaitingRequest(
-                LocalDate.of(2030, 8, 5), 6L, 10L);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(request)
-                .cookie(cookie)
-                .when().post("/waiting")
-                .then().log().all()
-                .statusCode(201)
-                .header("Location", "/waiting/1");
-
-        RestAssured.given().log().all()
-                .when().delete("/waiting/1")
-                .then().log().all()
-                .statusCode(204);
     }
 }
